# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DroniAuth API",
        "description": "OAuth 2.0 Authorization Server. Provides authentication, user management, and OAuth client management endpoints.",
        "contact": {
            "email": "dev@droni.co"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://auth.droni.co",
            "description": "Production"
        },
        {
            "url": "http://localhost",
            "description": "Local development"
        }
    ],
    "paths": {
...
        "/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Show profile edit page",
                "operationId": "8cfcc99d8d1aa366314cd8abf9cd9c9a",
                "responses": {
                    "200": {
                        "description": "Profile page (HTML/Inertia)"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sessionAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Profile"
                ],
                "summary": "Permanently delete the authenticated user's account",
                "operationId": "1d6a5c6489e35a763aa97ebef12863ce",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password"
                                ],
                                "properties": {
                                    "password": {
                                        "description": "Current password for confirmation",
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Account deleted – session invalidated, redirects to /"
                    },
                    "422": {
                        "description": "Wrong password"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sessionAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update profile name and/or email",
                "operationId": "354f5b23296bed4036564563e7423f43",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Jane Doe",
                                        "maxLength": 255
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jane@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Profile updated – redirects to /profile"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sessionAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sessionAuth": {
                "type": "apiKey",
                "description": "Session cookie (web authentication)",
                "name": "droniauth_session",
                "in": "cookie"
            },
            "oauth2": {
                "type": "oauth2",
                "description": "OAuth 2.0 Authorization Code flow issued by this server",
                "flows": {
                    "authorizationCode": {
                        "authorizationUrl": "/oauth/authorize",
                        "tokenUrl": "/oauth/token",
                        "refreshUrl": "/oauth/token/refresh",
                        "scopes": {
                            "profile": "Read user profile (name, avatar)",
                            "email": "Read email address",
                            "roles": "Read roles and permissions",
                            "admin": "Full admin access"
                        }
                    }
                }
            }
        }
    },
...
}